Created Pachinko Game v1
You can now play Pachinko. The camera rests in the center of 4 screens that are each cardinal direction around it. Click fire to shoot a Pachinko ball and if it hits anything on its way to a hole you earn points. After the last ball is destroyed, the game ends.

The mobile version has been optimized to run smoother. It does not like having over 450 pins to render and compute information for (understandably).